+++ pkg/proxy/integrations/mysqlparser/comPingPacket.go
@@ -1,0 +1,14 @@
+package mysqlparser
+
+import "errors"
+
+type ComPingPacket struct {
+}
+
+func decodeComPing(data []byte) (ComPingPacket, error) {
+	if len(data) < 1 || data[0] != 0x0e {
+		return ComPingPacket{}, errors.New("Data malformed for COM_PING")
+	}
+
+	return ComPingPacket{}, nil
+}
